+++ .gitignore
@@ -1,7 +1,2 @@
-twitter-2010.txt
-soc-LiveJournal1.txt
-soc-LiveJournal1.txt.bin
-dst.txt
-dst.bin
-twitter-2010.txt.bin
+inputs
 PageRankOutput.txt
