+++ cmake-build-debug/.ninja_log
@@ -1,12 +1,31 @@
 # ninja log v5
-1	384	1688208083836325390	CMakeFiles/untitled.dir/main.cpp.o	cf7b0c64d8a42b8d
-0	23	1688208299984298566	build.ninja	f76bddc10890553d
-0	265	1688208083716325391	CMakeFiles/untitled.dir/util.cpp.o	6ae0201750e3a3d9
-1	355	1688208083804325390	CMakeFiles/untitled.dir/EC.cpp.o	9fd08303b4513599
-0	307	1688208148276322193	CMakeFiles/untitled.dir/PageRank.cpp.o	b037a116bc86c60a
-307	369	1688208148336322188	untitled	97a26060c5fabc53
-1	261	1688208401824272422	CMakeFiles/untitled.dir/util.cpp.o	cdef0f33e5562385
-0	312	1688208401876272406	CMakeFiles/untitled.dir/PageRank.cpp.o	baedc2d68213391d
-0	364	1688208401928272391	CMakeFiles/untitled.dir/EC.cpp.o	3b5416954660cd8b
+0	655	1688578891014139955	CMakeFiles/main.cpp.dir/EC.cpp.o	7e8cccee35a145b1
 0	391	1688208401948272385	CMakeFiles/untitled.dir/main.cpp.o	6be52a0421b96607
+1	668	1688578891026139632	CMakeFiles/main.cpp.dir/ConvertToBinary.cpp.o	1e2b5d97cef48e45
+0	472	1688584060483718192	CMakeFiles/ConvertToBinary.dir/graph-creation/parallelEC.cpp.o	60fb0eee1c12a05c
+0	434	1688584060443718574	CMakeFiles/ConvertToBinary.dir/utils/ConvertToBinary.cpp.o	a901deff47fc733b
+0	23	1688584651971187194	build.ninja	f76bddc10890553d
+0	467	1688584032999996839	CMakeFiles/gpf.dir/utils/util.cpp.o	21c56e729a8223e9
+1	261	1688208401824272422	CMakeFiles/untitled.dir/util.cpp.o	cdef0f33e5562385
+0	364	1688208401928272391	CMakeFiles/untitled.dir/EC.cpp.o	3b5416954660cd8b
+0	614	1688578890974141034	CMakeFiles/main.cpp.dir/PageRank.cpp.o	95cb5f8117d0e59f
+1	320	1688579025959320698	CMakeFiles/ConvertToBinary.dir/ConvertToBinary.cpp.o	2251f8ac03f233f
+0	548	1688583887070077557	CMakeFiles/main.cpp.dir/graph-creation/EC.cpp.o	b1b50690e6db27eb
+0	557	1688584033087995896	CMakeFiles/gpf.dir/graph-creation/EC.cpp.o	6fadf526ce1b73e
+1	550	1688578890910142762	CMakeFiles/main.cpp.dir/util.cpp.o	9b24f2e890c09d7c
+0	312	1688208401876272406	CMakeFiles/untitled.dir/PageRank.cpp.o	baedc2d68213391d
+0	554	1688584033087995896	CMakeFiles/gpf.dir/main.cpp.o	688963e4e78c2e4
+1	558	1688583887082077340	CMakeFiles/main.cpp.dir/main.cpp.o	7da4ce6ed935758d
 391	451	1688208402016272365	untitled	730207d2a5becdd3
+0	482	1688583887006078718	CMakeFiles/main.cpp.dir/utils/util.cpp.o	2bb615759a1740dc
+1	475	1688584667915268782	CMakeFiles/gpf.dir/utils/ConvertToBinary.cpp.o	add34f3bc8c80760
+1	509	1688584667947268948	CMakeFiles/gpf.dir/algorithms/PageRank.cpp.o	de11b5be3cc2a2f8
+1	556	1688584667995269198	CMakeFiles/gpf.dir/graph-creation/parallelEC.cpp.o	973dd37925f18d8d
+1	318	1688584772875881887	CMakeFiles/gpf.dir/utils/ConvertToBinary.cpp.o	add34f3bc8c80760
+1	363	1688584772919882169	CMakeFiles/gpf.dir/main.cpp.o	688963e4e78c2e4
+0	373	1688584839560332130	CMakeFiles/gpf.dir/graph-creation/parallelEC.cpp.o	973dd37925f18d8d
+1	372	1688584859364473855	CMakeFiles/gpf.dir/graph-creation/parallelEC.cpp.o	973dd37925f18d8d
+1	327	1688584885304664641	CMakeFiles/gpf.dir/graph-creation/parallelEC.cpp.o	973dd37925f18d8d
+327	414	1688584885388665268	gpf	3e9408a45a2eb12a
+1	370	1688584957793226482	CMakeFiles/gpf.dir/main.cpp.o	688963e4e78c2e4
+370	458	1688584957881227188	gpf	3e9408a45a2eb12a
