+++ cmake-build-debug/CMakeCache.txt
@@ -148,7 +148,7 @@ CMAKE_PROJECT_DESCRIPTION:STATIC=
 CMAKE_PROJECT_HOMEPAGE_URL:STATIC=
 
 //Value Computed by CMake
-CMAKE_PROJECT_NAME:STATIC=untitled
+CMAKE_PROJECT_NAME:STATIC=graph_processing_framework
 
 //Path to a program.
 CMAKE_RANLIB:FILEPATH=/usr/bin/ranlib
@@ -211,6 +211,42 @@ CMAKE_STRIP:FILEPATH=/usr/bin/strip
 // during the make.  This is useful for debugging only. With Visual
 // Studio IDE projects all commands are done without /nologo.
 CMAKE_VERBOSE_MAKEFILE:BOOL=FALSE
+
+//Value Computed by CMake
+graph_processing_BINARY_DIR:STATIC=/home/farouk/CLionProjects/untitled/cmake-build-debug
+
+//Value Computed by CMake
+graph_processing_IS_TOP_LEVEL:STATIC=ON
+
+//Value Computed by CMake
+graph_processing_SOURCE_DIR:STATIC=/home/farouk/CLionProjects/untitled
+
+//Value Computed by CMake
+graph_processing__BINARY_DIR:STATIC=/home/farouk/CLionProjects/untitled/cmake-build-debug
+
+//Value Computed by CMake
+graph_processing__IS_TOP_LEVEL:STATIC=ON
+
+//Value Computed by CMake
+graph_processing__SOURCE_DIR:STATIC=/home/farouk/CLionProjects/untitled
+
+//Value Computed by CMake
+graph_processing_frame_BINARY_DIR:STATIC=/home/farouk/CLionProjects/untitled/cmake-build-debug
+
+//Value Computed by CMake
+graph_processing_frame_IS_TOP_LEVEL:STATIC=ON
+
+//Value Computed by CMake
+graph_processing_frame_SOURCE_DIR:STATIC=/home/farouk/CLionProjects/untitled
+
+//Value Computed by CMake
+graph_processing_framework_BINARY_DIR:STATIC=/home/farouk/CLionProjects/untitled/cmake-build-debug
+
+//Value Computed by CMake
+graph_processing_framework_IS_TOP_LEVEL:STATIC=ON
+
+//Value Computed by CMake
+graph_processing_framework_SOURCE_DIR:STATIC=/home/farouk/CLionProjects/untitled
 
 //Value Computed by CMake
 untitled_BINARY_DIR:STATIC=/home/farouk/CLionProjects/untitled/cmake-build-debug
