+++ cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,3 @@
-/home/farouk/CLionProjects/untitled/cmake-build-debug/CMakeFiles/untitled.dir
+/home/farouk/CLionProjects/untitled/cmake-build-debug/CMakeFiles/gpf.dir
 /home/farouk/CLionProjects/untitled/cmake-build-debug/CMakeFiles/edit_cache.dir
 /home/farouk/CLionProjects/untitled/cmake-build-debug/CMakeFiles/rebuild_cache.dir
