+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Jul 01 11:46 CET
+Start testing: Jul 05 20:22 CET
 ----------------------------------------------------------
-End testing: Jul 01 11:46 CET
+End testing: Jul 05 20:22 CET
