+++ cmake-build-debug/soc-LiveJournal1.txt.conf
@@ -1,0 +1,1 @@
+1
