--- dst.bin
@@ -1,5 +1,0 @@
-00000000: 00000001 00000000 00000000 00000000 00000010 00000000  ......
-00000006: 00000000 00000000 00000011 00000000 00000000 00000000  ......
-0000000c: 00000100 00000000 00000000 00000000 00000101 00000000  ......
-00000012: 00000000 00000000 00000110 00000000 00000000 00000000  ......
-00000018: 00000100 00000000 00000000 00000000                    ....
