+++ utils/ConvertToBinary.h
@@ -1,0 +1,11 @@
+//
+// Created by farouk on 05/07/23.
+//
+#include <iostream>
+
+#ifndef GRAPH_PROCESSING_FRAMEWORK_CONVERTTOBINARY_H
+#define GRAPH_PROCESSING_FRAMEWORK_CONVERTTOBINARY_H
+
+void convertToBinary(const std::string& path);
+
+#endif //GRAPH_PROCESSING_FRAMEWORK_CONVERTTOBINARY_H
